feat(homeTableColumnManagementMenu): Renamed folder to HomeTableState, created separate file to hold function UpdateColumnManagement.js
