+++ src/Analysis/GWASResults/Views/Home/HomeTableState/DetermineInitialColumnManagement.js
