+++ src/Analysis/GWASResults/Views/Home/HomeTableState/InitialHomeTableState.js
