+++ src/Analysis/GWASResults/Views/Home/HomeTableState/UpdateColumnManagement.js
@@ -1,0 +1,10 @@
+const UpdateColumnManagement = (homeTableState) => {
+  if (homeTableState.useLocalStorage) {
+    localStorage.setItem(
+      'columnManagement',
+      JSON.stringify(homeTableState.columnManagement)
+    );
+  }
+};
+
+export default UpdateColumnManagement;
