+++ src/Analysis/GWASResults/Views/Home/HomeTableState/localStorageAvailable.js
